feat(events): Add eventPattern to settings
Event Patterns allow you to restrict the routes that the CORS module will process.
By default, the module processes all routes.  The setting can take either
a single event pattern as a string or an array of event patterns. If any
event pattern matches, the request will be processed through the CORS interceptor.